Desktop/mcp-servers/outlook-mcp-servers/requirements.txt
fastapi==0.131.0
uvicorn[standard]==0.41.0

mcp==1.26.0

httpx==0.28.1
msal==1.33.0

pydantic-settings==2.13.1
python-dotenv==1.0.1